src/requirements.txt
pytest~=7.4.4
pydantic~=2.5.3
pydantic[email]~=2.5.3
bcrypt~=4.0.1
Flask~=3.0.0
gunicorn~=20.1.0
PyJWT~=2.3.0
DateTime~=5.4
redis~=3.5.3
flask-SQLAlchemy==3.1.1
python-dotenv~=1.0.0
environs~=10.3.0
psycopg2-binary~=2.9.9
googlemaps==4.10.0
Flask-Pydantic==0.12.0
coverage==7.4.3 
pytest-cov==4.1.0
segno==1.6.1
